♻️  Move functions to noir_utils file
