+++ src/utils/mod.rs
@@ -5,5 +5,6 @@ pub mod crypto;
 pub mod db;
 pub mod logger;
 pub mod modules;
+pub mod noir_utils;
 pub mod serde;
 pub mod static_type_map;
